+++ src/editorOps.h
@@ -1,0 +1,8 @@
+#ifndef EDITOR_OPS_C
+#define EDITOR_OPS_C
+
+void editorInsertChar(int c);
+void editorInsertNewline();
+void editorDelChar();
+
+#endif
