+++ src/fileIO.c
@@ -1,0 +1,107 @@
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <fcntl.h>
+#include <unistd.h>
+#include "tedit.h"
+#include "rowOps.h"
+
+char *editorRowsToString(int *buflen)
+{
+	int totlen = 0;
+	int j;
+	for (j = 0; j < E.numrows; j++)
+		totlen += E.row[j].size + 1;
+	*buflen = totlen;
+
+	char *buf = malloc(totlen);
+	char *p = buf;
+	for (j = 0; j < E.numrows; j++)
+	{
+		memcpy(p, E.row[j].chars, E.row[j].size);
+		p += E.row[j].size;
+		*p = '\n';
+		p++;
+	}
+
+	return buf;
+}
+
+void editorOpen(char *filename)
+{
+	free(E.filename);
+	E.filename = strdup(filename);
+
+	editorSelectSyntaxHighlight();
+
+	FILE *fp = fopen(filename, "r");
+	if (!fp) die ("fopen");
+	
+	char *line = NULL;
+	size_t linecap = 0;
+	ssize_t linelen;
+
+	while ((linelen = getline(&line, &linecap, fp)) != -1)
+	{
+		while (linelen > 0 && (line[linelen - 1] == '\n' ||
+								line[linelen - 1] == '\r'))
+			linelen--;
+		editorInsertRow(E.numrows, line, linelen);	
+	}
+	free(line);
+	fclose(fp);	
+	E.dirty = 0;
+}
+
+void editorOpenFromEdit()
+{
+	char *file;
+	file = editorPrompt("Open file: %s (ESC to cancel)", NULL);
+	if (file == NULL) {
+		editorSetStatusMessage("Open aborted");
+		return;
+	}
+	initEditor();
+	editorOpen(file);
+}
+
+// TODO: This should in the future create a temporary file and
+//       write to that, then renaming the temp file to the file
+//       name that we want to write to but for now this works.
+void editorSave()
+{
+	if (E.filename == NULL) 
+	{
+		E.filename = editorPrompt("Save as: %s (ESC to cancel)", NULL);
+		if (E.filename == NULL) {
+			editorSetStatusMessage("Save aborted");
+			return;
+		}
+		editorSelectSyntaxHighlight();
+	}
+
+	int len;
+	char *buf = editorRowsToString(&len);
+
+	int fd = open(E.filename, O_RDWR | O_CREAT, 0644);
+	if (fd != -1)
+	{
+		if (ftruncate(fd, len) != -1)
+		{
+			if (write(fd, buf, len) == len)
+			{
+				close(fd);
+				free(buf);
+				E.dirty = 0;
+				editorSetStatusMessage("%d bytes written to disk", len);
+				return;
+			}
+		}
+		close(fd);
+	}
+	free(buf);
+	editorSetStatusMessage("Can't save! I/O error: %s", strerror(errno));
+}
