+++ src/fileIO.h
@@ -1,0 +1,9 @@
+#ifndef FILE_IO_C_
+#define FILE_IO_C_
+
+char *editorRowsToString(int *buflen);
+void editorOpen(char *filename);
+void editorOpenFromEdit();
+void editorSave();
+
+#endif
