+++ src/rowOps.h
@@ -1,0 +1,16 @@
+#ifndef ROW_OPS_C_
+#define ROW_OPS_C_
+
+#include "tedit.h"
+
+int editorRowCxToRx(erow *row, int cx);
+int editorRowRxToCx(erow *row, int rx);
+void editorUpdateRow(erow *row);
+void editorInsertRow(int at, char *s, size_t len);
+void editorFreeRow(erow *row);
+void editorDelRow(int at);
+void editorRowInsertChar(erow *row, int at, int c);
+void editorRowAppendString(erow *row, char *s, size_t len);
+void editorRowDelChar(erow *row, int at);
+
+#endif
