+++ src/tedit.c
@@ -11,94 +11,15 @@
 #include <stdarg.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/ioctl.h>
 #include <sys/types.h>
 #include <termios.h>
 #include <time.h>
 #include <unistd.h>
-
-// Defines
-#define TEDIT_VERSION "1.0.3"
-#define TEDIT_TAB_STOP 8
-#define TEDIT_QUIT_TIMES 3
-#define TEDIT_OPEN_TIMES 3
-
-#define CTRL_KEY(k) ((k) & 0x1f)
-
-enum editorKey {
-	BACKSPACE = 127,
-	ARROW_LEFT = 1000,
-	ARROW_RIGHT,
-	ARROW_UP,
-	ARROW_DOWN,
-	DEL_KEY,
-	HOME_KEY,
-	END_KEY,
-	PAGE_UP,
-	PAGE_DOWN
-};
-
-enum editorHighlight {
-	HL_NORMAL = 0,
-	HL_COMMENT,
-	HL_MLCOMMENT,
-	HL_KEYWORD1,
-	HL_KEYWORD2,
-	HL_STRING,
-	HL_NUMBER,
-	HL_MATCH
-};
-
-#define HL_HIGHLIGHT_NUMERS (1<<0)
-#define HL_HIGHLIGHT_STRINGS (1<<1)
-
-// DATA
-
-struct editorSyntax {
-	char *filetype;
-	char **filematch;
-	char **keywords;
-	char *singleline_comment_start;
-	char *multiline_comment_start;
-	char *multiline_comment_end;
-	int flags;
-};
-
-typedef struct erow 
-{
-	int idx;
-	int size;
-	int rsize;
-	char *chars;
-	char *render;
-	unsigned char *hl;
-	int hl_open_comment;
-} erow;
-
-struct editorConfig 
-{
-	int cx, cy; // Cursor x and y position
-	int rx;
-
-	int rowoff;
-	int coloff;
-
-	int screenrows;
-	int screencols;
-
-	int numrows;
-	erow *row;
-	int dirty;
-
-	char *filename;
-	char statusmsg[80];
-	time_t statusmsg_time;
-
-	struct editorSyntax *syntax;
-	struct termios orig_termios;
-};
-
-struct editorConfig E;
+#include "editorOps.h"
+#include "tedit.h"
+#include "fileIO.h"
+#include "canonical.h"
+#include "rowOps.h"
 
 // FIletypes
 
@@ -120,14 +41,6 @@ struct editorSyntax HLDB[] = {
 };
 #define HLDB_ENTRIES (sizeof(HLDB) / sizeof(HLDB[0]))
 
-// Prototyes
-void initEditor();
-void editorSetStatusMessage(const char *fmt, ...);
-void editorRefreshScreen();
-char *editorPrompt(char *prompt, void(*callback)(char *, int));
-
-// Terminal
-
 
 // Print error message and exit the program
 void die(const char *s)
@@ -137,152 +50,6 @@ void die(const char *s)
 
 	perror(s);
 	exit(1);
-}
-
-/*
- *  Canonical mode, or cooked mode makes it so that every character that 
- *  is entered gets echoed. We obviously don't want this for a text editor, 
- *  so we must enter raw mode, which makes the terminal not echo everything
- *  that's typed. When the script starts it should enter raw mode, and when
- *  it ends it should exit so the user isn't stuck in raw mode.
- */
-
-void disableRawMode()
-{
-	if (tcsetattr(STDIN_FILENO, TCSAFLUSH, &E.orig_termios) == -1)
-		die("tcsetattr");
-}
-
-void enableRawMode() 
-{
-	if (tcgetattr(STDIN_FILENO, &E.orig_termios) == -1) die("tcgetattr");
-	atexit(disableRawMode);
-
-	struct termios raw = E.orig_termios;
-
-			// Unknown is various antiquated flags to turn off
-	raw.c_iflag &= ~(BRKINT | ICRNL | INPCK | ISTRIP | IXON); 		
-			// Disable unknown, ctrl-m, unknown, unknown, and 
-			// ctrl-s, ctrl-q (software flow control)
-	raw.c_oflag &= ~(OPOST); 
-			// Disable automatic conversion from \n to \r\n
-	raw.c_cflag |= (CS8);
-			// Unknown
-	raw.c_lflag &= ~(ECHO | ICANON | IEXTEN | ISIG); 
-			// Flips the bits for echo, canonical mode, ctrl-v,
-			// and interrupt signal (ctrl-c, ctrl-z)
-	
-	// Set timeout for read() so that it returns if if doesn't get any
-	// input for a certain amount of time
-	raw.c_cc[VMIN] = 0;  // Set minimum number of bytes is needed before
-			     // read() can return. 0 Means that it can return
-			     // instantly
-	raw.c_cc[VTIME] = 1; // Set maximum amount of time to wait before 
-			     // read() returns in tenths of seconds.
-
-	tcsetattr(STDIN_FILENO, TCSAFLUSH, &raw);
-
-	if (tcsetattr(STDIN_FILENO, TCSAFLUSH, &raw) == -1) die("tcsetattr");
-}
-
-int editorReadKey()
-{
-	int nread;
-	char c;
-
-	while ((nread = read(STDIN_FILENO, &c, 1)) != 1)
-	{
-		if (nread == -1 && errno != EAGAIN) die("read");
-	}
-
-	if (c == '\x1b')
-	{
-		char seq[3];
-
-		if (read(STDIN_FILENO, &seq[0], 1) != 1) return '\x1b';
-		if (read(STDIN_FILENO, &seq[1], 1) != 1) return '\x1b';
-
-		if (seq[0] == '[')
-		{
-			if (seq[1] >= '0' && seq[1] <= '9') 
-			{
-				if (read(STDIN_FILENO, &seq[2], 1) != 1) return '\x1b';
-				if (seq[2] == '~') 
-				{
-					switch (seq[1])
-					{
-						case '1': return HOME_KEY;
-						case '3': return DEL_KEY;
-						case '4': return END_KEY;
-						case '5': return PAGE_UP;
-						case '6': return PAGE_DOWN;
-						case '7': return HOME_KEY;
-						case '8': return END_KEY;
-					}
-				}
-			} else {
-				switch (seq[1]) 
-				{
-					case 'A': return ARROW_UP;
-					case 'B': return ARROW_DOWN;
-					case 'C': return ARROW_RIGHT;
-					case 'D': return ARROW_LEFT;
-					case 'H': return HOME_KEY;
-					case 'F': return END_KEY;
-				}
-			}
-		} else if (seq[0] == 'O') {
-			switch (seq[1]) 
-			{
-				case 'H': return HOME_KEY;
-				case 'F': return END_KEY;
-			}
-		}
-
-		return '\x1b';
-	} else {
-		return c;
-	}
-
-	return c;
-}
-
-int getCursorPosition(int *rows, int *cols)
-{
-	char buf[32];
-	unsigned int i = 0;
-	
-	if (write(STDOUT_FILENO, "\x1b[6n", 4) != 4) return -1;
-
-	printf("\r\n");
-	
-	while (i < sizeof(buf) - 1)
-	{
-		if (read(STDIN_FILENO, &buf[i], 1) != 1) break;
-		if (buf[i] == 'R') break;
-		i++;
-	}
-	buf[i] = '\0';
-
-	if (buf[0] != '\x1b' || buf[1] != '[') return -1;
-	if (sscanf(&buf[2], "%d;%d", rows, cols) != 2) return -1;
-
-	return 0;
-}
-
-int getWindowSize(int *rows, int *cols)
-{
-	struct winsize ws;
-
-	if (ioctl(STDOUT_FILENO, TIOCGWINSZ, &ws) == -1 || ws.ws_col == 0)
-	{
-		if (write(STDOUT_FILENO, "\x1b[999C\x1b[999B", 12) != 12) return -1;
-		return getCursorPosition(rows, cols);
-	} else {
-		*cols = ws.ws_col;
-		*rows = ws.ws_row;
-		return 0;
-	}
 }
 
 // Syntax highlighting
@@ -475,282 +242,9 @@ void editorSelectSyntaxHighlight()
 	}
 }
 
-// Row Operations
 
-int editorRowCxToRx(erow *row, int cx)
-{
-	int rx = 0;
-	int j;
-	for (j = 0; j< cx; j++)
-	{
-		if(row->chars[j] == '\t')
-			rx += (TEDIT_TAB_STOP - 1) - (rx % TEDIT_TAB_STOP);
-		rx++;
-	}
-	return rx;
-}
 
-int editorRowRxToCx(erow *row, int rx)
-{
-	int cur_rx = 0;
-	int cx;
-	for (cx = 0; cx < row->size; cx++) 
-	{
-		if (row->chars[cx] == '\t')
-			cur_rx += (TEDIT_TAB_STOP - 1) - (cur_rx % TEDIT_TAB_STOP);
-		cur_rx++;
-		if (cur_rx > rx) return cx;
-	}
-	return cx;
-}
 
-void editorUpdateRow(erow *row)
-{
-	int tabs = 0;
-	int j;
-	for (j = 0; j< row->size; j++)
-		if(row->chars[j] == '\t') tabs++;
-
-	free(row->render);
-	row->render = malloc(row->size + tabs*(TEDIT_TAB_STOP -1) + 1);
-
-	int idx = 0;
-	for (j = 0; j < row->size; j++)
-	{
-		if(row->chars[j] == '\t')
-		{
-			row->render[idx++] = ' ';
-			while (idx % TEDIT_TAB_STOP != 0) row->render[idx++] = ' ';
-		} else {
-			row->render[idx++] = row->chars[j];
-		}
-	}
-	row->render[idx] = '\0';
-	row->rsize = idx;
-
-	editorUpdateSyntax(row);
-}
-
-void editorInsertRow(int at, char *s, size_t len)
-{
-	if (at < 0 || at > E.numrows) return;
-	
-	E.row = realloc(E.row, sizeof(erow) * (E.numrows + 1));
-	memmove(&E.row[at + 1], &E.row[at], sizeof(erow) * (E.numrows - at));
-	for (int j = at + 1; j<=E.numrows; j++) E.row[j].idx++;
-
-	E.row[at].idx = at;
-
-	E.row[at].size = len;
-	E.row[at].chars = malloc(len + 1);
-	memcpy(E.row[at].chars, s, len);
-	E.row[at].chars[len] = '\0';
-
-	E.row[at].rsize = 0;
-	E.row[at].render = NULL;
-	E.row[at].hl = NULL;
-	E.row[at].hl_open_comment = 0;
-	editorUpdateRow(&E.row[at]);
-
-	E.numrows++;
-	E.dirty++;
-}
-
-void editorFreeRow(erow *row)
-{
-	free(row->render);
-	free(row->chars);
-	free(row->hl);
-}
-
-void editorDelRow(int at)
-{
-	if (at < 0 || at >= E.numrows) return;
-	editorFreeRow(&E.row[at]);
-	memmove(&E.row[at], &E.row[at + 1], sizeof(erow) * (E.numrows - at - 1));
-	for (int j = at; j<E.numrows - 1; j++) E.row[j].idx--;
-	E.numrows--;
-	E.dirty++;
-}
-
-void editorRowInsertChar(erow *row, int at, int c)
-{
-	if (at < 0 || at > row->size) at = row->size;
-	row->chars = realloc(row->chars, row->size + 2);
-	memmove(&row->chars[at + 1], &row->chars[at], row->size - at + 1);
-	row->size++;
-	row->chars[at] = c;
-	editorUpdateRow(row);
-	E.dirty++;
-}
-
-void editorRowAppendString(erow *row, char *s, size_t len)
-{
-	row->chars = realloc(row->chars, row->size + len + 1);
-	memcpy(&row->chars[row->size], s, len);
-	row->size += len;
-	row->chars[row->size] = '\0';
-	editorUpdateRow(row);
-	E.dirty++;
-}
-
-void editorRowDelChar(erow *row, int at)
-{
-	if (at < 0 || at >= row->size) return;
-	memmove(&row->chars[at], &row->chars[at + 1], row->size - at);
-	row->size--;
-	editorUpdateRow(row);
-	E.dirty++;
-}
-
-// Editor operations
-
-void editorInsertChar(int c)
-{
-	if (E.cy == E.numrows)
-	{
-		editorInsertRow(E.numrows, "", 0);
-	}
-	editorRowInsertChar(&E.row[E.cy], E.cx, c);
-	E.cx++;
-}
-
-void editorInsertNewline()
-{
-	if (E.cx == 0)
-	{
-		editorInsertRow(E.cy, "", 0);
-	} else {
-		erow *row = &E.row[E.cy];
-		editorInsertRow(E.cy + 1, &row->chars[E.cx], row->size - E.cx);
-		row = &E.row[E.cy];
-		row->size = E.cx;
-		row->chars[row->size] = '\0';
-		editorUpdateRow(row);
-	}
-	E.cy++;
-	E.cx = 0;
-}
-
-void editorDelChar()
-{
-	if (E.cy == E.numrows) return;
-	if (E.cx == 0 && E.cy == 0) return;
-
-	erow *row = &E.row[E.cy];
-	if (E.cx > 0)
-	{
-		editorRowDelChar(row, E.cx - 1);
-		E.cx--;
-	} else {
-		E.cx = E.row[E.cy - 1].size;
-		editorRowAppendString(&E.row[E.cy - 1], row->chars, row->size);
-		editorDelRow(E.cy);
-		E.cy--;
-	}
-}
-
-// File I/O
-
-char *editorRowsToString(int *buflen)
-{
-	int totlen = 0;
-	int j;
-	for (j = 0; j < E.numrows; j++)
-		totlen += E.row[j].size + 1;
-	*buflen = totlen;
-
-	char *buf = malloc(totlen);
-	char *p = buf;
-	for (j = 0; j < E.numrows; j++)
-	{
-		memcpy(p, E.row[j].chars, E.row[j].size);
-		p += E.row[j].size;
-		*p = '\n';
-		p++;
-	}
-
-	return buf;
-}
-
-void editorOpen(char *filename)
-{
-	free(E.filename);
-	E.filename = strdup(filename);
-
-	editorSelectSyntaxHighlight();
-
-	FILE *fp = fopen(filename, "r");
-	if (!fp) die ("fopen");
-	
-	char *line = NULL;
-	size_t linecap = 0;
-	ssize_t linelen;
-
-	while ((linelen = getline(&line, &linecap, fp)) != -1)
-	{
-		while (linelen > 0 && (line[linelen - 1] == '\n' ||
-								line[linelen - 1] == '\r'))
-			linelen--;
-		editorInsertRow(E.numrows, line, linelen);	
-	}
-	free(line);
-	fclose(fp);	
-	E.dirty = 0;
-}
-
-void editorOpenFromEdit()
-{
-	// I don't know why it's not letting us use E.filename instead of
-	// *file right now, but it's not working. fopen will throw a 
-	// file not found error. This works tho
-	char *file;
-	file = editorPrompt("Open file: %s (ESC to cancel)", NULL);
-	if (file == NULL) {
-		editorSetStatusMessage("Open aborted");
-		return;
-	}
-	initEditor();
-	editorOpen(file);
-}
-
-// TODO: This should in the future create a temporary file and
-//       write to that, then renaming the temp file to the file
-//       name that we want to write to but for now this works.
-void editorSave()
-{
-	if (E.filename == NULL) 
-	{
-		E.filename = editorPrompt("Save as: %s (ESC to cancel)", NULL);
-		if (E.filename == NULL) {
-			editorSetStatusMessage("Save aborted");
-			return;
-		}
-		editorSelectSyntaxHighlight();
-	}
-
-	int len;
-	char *buf = editorRowsToString(&len);
-
-	int fd = open(E.filename, O_RDWR | O_CREAT, 0644);
-	if (fd != -1)
-	{
-		if (ftruncate(fd, len) != -1)
-		{
-			if (write(fd, buf, len) == len)
-			{
-				close(fd);
-				free(buf);
-				E.dirty = 0;
-				editorSetStatusMessage("%d bytes written to disk", len);
-				return;
-			}
-		}
-		close(fd);
-	}
-	free(buf);
-	editorSetStatusMessage("Can't save! I/O error: %s", strerror(errno));
-}
 
 // Find
 void editorFindCallback(char *query, int key)
@@ -1231,7 +725,7 @@ void initEditor()
 
 int main(int argc, char *argv[]) 
 {
-	enableRawMode(); 
+	enableRawMode(E); 
 	initEditor();
 	if (argc >= 2)
 	{
